reference/sam/deprecated/spotify/requirements.txt
# spotipy
python-dotenv
pydash
#https://github.com/supabase-community/supabase-py
supabase-py #supabase
# pocketbase
# redislite